chara_fur: makeFurTex rootX/rootZ rand conversion is signed int->float
Reloc-value census: target loads kYmEnvSignedDoubleMagic (0x4330000080000000)
with xoris for the FurRandNext() result conversion in the hair loop; ours
emitted the unsigned bias (@1135 = 0x4330000000000000, no xoris). Value
verified against orig main.dol bytes. Behavior identical (rand % 32767 is
always non-negative). makeFurTex 76.118 -> 76.413, unit 79.799 -> 79.870.

+++ src/chara_fur.cpp
@@ -2155,8 +2155,8 @@ void CChara::makeFurTex()
 		for (unsigned int hair = 0; hair < 0x4000; hair++) {
 			GXBegin(GX_LINESTRIP, GX_VTXFMT0, 5);
 
-			float rootX = kCharaFurDepthScaleBase * (FLOAT_80331164 * static_cast<float>(FurRandNext()) + kCharaFurViewDepthThreshold);
-			float rootZ = kCharaFurDepthScaleBase * (FLOAT_80331164 * static_cast<float>(FurRandNext()) + kCharaFurViewDepthThreshold);
+			float rootX = kCharaFurDepthScaleBase * (FLOAT_80331164 * static_cast<float>(static_cast<int>(FurRandNext())) + kCharaFurViewDepthThreshold);
+			float rootZ = kCharaFurDepthScaleBase * (FLOAT_80331164 * static_cast<float>(static_cast<int>(FurRandNext())) + kCharaFurViewDepthThreshold);
 			CVector root(rootX, kCharaFurDepthZero, rootZ);
 
 			s_mogFurRand = s_mogFurRand * 0x41C64E6D + 0x3039;
